requirements.txt
six==1.12.0
websocket-client==0.56.0
six==1.12.0
websocket-client==0.48.0
